HSO: Minor changes caused by Bluetooth and motor tests; do not work perfect yet.

+++ examples/CarControlSystem/application/src/carcontrol/device/Engine.java
@@ -5,8 +5,8 @@ public interface Engine {
 	void engineOn();
 	void engineOff();
 	
-	byte getEngineSpeed();
-	void setEngineSpeed (byte speedPercent); 
-	void setEngineBrake (byte brakePercent);
+	int getEngineSpeed();
+	void setEngineSpeed (int speedPercent); 
+	void setEngineBrake (int brakePercent);
 	
 }
